feat: change dashbord layout and optimize style

+++ client/src/components/leadsTable/DeleteRow.tsx
@@ -1,6 +1,6 @@
 import { Lead } from "../../types";
 import { useDeleteLead } from "../../features/lead/hooks";
-import { TiDelete } from "react-icons/ti";
+import { LuTrash2 } from "react-icons/lu";
 
 type Props = {
   id: Lead["id"];
@@ -10,12 +10,17 @@ const DeleteRow = ({ id }: Props) => {
   const deleteLead = useDeleteLead();
 
   const handleDelete = async () => {
-    await deleteLead(id);
+    const confirmed = window.confirm(
+      "Are you sure you want to delete this lead? This action cannot be undone."
+    );
+    if (confirmed) {
+      await deleteLead(id);
+    }
   };
 
   return (
     <button onClick={handleDelete} className="text-red-600 hover:text-red-900">
-      <TiDelete size={27} />
+      <LuTrash2 size={23} />
     </button>
   );
 };
